requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-aiml==0.9.3
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
joblib==1.0.1
plotly==5.3.1
python-multipart==0.0.5
jinja2==3.0.1
aiofiles==0.7.0
websockets==10.0
python-dotenv==0.19.0